Refactor DateTime util funcs; add parseTime func
